requirements.txt
fastapi
uvicorn
openai-whisper
torch
python-multipart
ffmpeg-python
